refactor: implement bottomsheet databinding
