+++ app/src/main/java/campus/tech/kakao/map/utill/BitmapUtils.kt
@@ -1,0 +1,24 @@
+package campus.tech.kakao.map.utill
+
+import android.content.Context
+import android.graphics.Bitmap
+import android.graphics.Canvas
+import androidx.core.content.ContextCompat
+
+class BitmapUtils {
+
+    companion object {
+        fun vectorToBitmap(context: Context, drawableId: Int): Bitmap {
+            val drawable = ContextCompat.getDrawable(context, drawableId)
+            val bitmap = Bitmap.createBitmap(
+                drawable!!.intrinsicWidth,
+                drawable!!.intrinsicHeight,
+                Bitmap.Config.ARGB_8888
+            )
+            val canvas = Canvas(bitmap)
+            drawable.setBounds(0, 0, canvas.width, canvas.height)
+            drawable.draw(canvas)
+            return bitmap
+        }
+    }
+}
