+++ app/src/main/res/layout/activity_main.xml
@@ -1,4 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
+
 <androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
@@ -9,13 +10,13 @@
 
     <androidx.cardview.widget.CardView
         android:id="@+id/search_window"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        android:layout_margin="20dp"
         android:layout_width="match_parent"
         android:layout_height="50dp"
+        android:layout_margin="20dp"
         app:cardCornerRadius="5dp"
-        app:cardElevation="8dp">
+        app:cardElevation="8dp"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent">
 
         <LinearLayout
             android:layout_width="match_parent"
@@ -27,12 +28,12 @@
 
             <EditText
                 android:id="@+id/search_input"
-                android:focusable="false"
-                android:focusableInTouchMode="false"
-                android:layout_weight="1"
                 android:layout_width="0dp"
                 android:layout_height="wrap_content"
+                android:layout_weight="1"
                 android:background="@android:color/transparent"
+                android:focusable="false"
+                android:focusableInTouchMode="false"
                 android:hint="@string/search_input_hint" />
 
             <ImageView
